32-bit build fix

+++ src/structwrappers.rs
@@ -907,7 +907,7 @@ impl DescriptorHeap {
     pub fn cpu_handle(&self, i: u32) -> D3D12_CPU_DESCRIPTOR_HANDLE {
         assert!(i < self.heap_size);
         let mut handle = self.dheap.get_cpu_descriptor_handle_for_heap_start();
-        handle.ptr += (i as SIZE_T) * self.handle_size;
+        handle.ptr += (i as u64) * (self.handle_size as u64);
         handle
     }
